Consolidate isOriginOk check
Merge the two copies in routes/my_api and routes/thingpedia_api,
and explain what's going on.
